requirements.txt
google-cloud-pubsub
google-cloud-core
sh
PyInstaller